# src/main/resources/openapi/media-api-contract.yaml
openapi: 3.0.3
info:
  title: VaultApp Media API
  description: REST API for managing different types of media (manga, series, films, comics, books, anime)
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
paths:
...
  /comics:
    get:
      summary: Get all comics
      description: Retrieve a list of all comics in the database, with optional pagination.
      operationId: getAllComics
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size for pagination
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of comics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComicsDTO'
        '500':
          description: Internal server error
    post:
      summary: Create a new comic
      description: Adds a new comic to the database.
      operationId: createComics
      requestBody:
        description: Comics object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComicsDTO'
      responses:
        '201':
          description: Comic created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicsDTO'
        '400':
          description: Invalid comics data
        '500':
          description: Internal server error
  /comics/{id}:
    get:
      summary: Get a specific comic by ID
      description: Retrieve a specific comic by its unique identifier.
      operationId: getComicsById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the comic to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: The comic object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicsDTO'
        '404':
          description: Comic not found
        '500':
          description: Internal server error
    put:
      summary: Update a comic
      description: Update the details of an existing comic.
      operationId: updateComics
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the comic to update
          schema:
            type: integer
      requestBody:
        description: Updated comics object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComicsDTO'
      responses:
        '200':
          description: Comic updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicsDTO'
        '400':
          description: Invalid comics data
        '404':
          description: Comic not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a comic
      description: Deletes a comic by its unique identifier.
      operationId: deleteComics
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the comic to delete
          schema:
            type: integer
      responses:
        '204':
          description: Comic deleted successfully
        '404':
          description: Comic not found
        '500':
          description: Internal server error
  /book:
    get:
      summary: Get all books
      description: Retrieve a list of all books in the database, with optional pagination.
      operationId: getAllBooks
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size for pagination
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookDTO'
        '500':
          description: Internal server error
    post:
      summary: Create a new book
      description: Adds a new book to the database.
      operationId: createBook
      requestBody:
        description: Book object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookDTO'
      responses:
        '201':
          description: Book created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDTO'
        '400':
          description: Invalid book data
        '500':
          description: Internal server error
  /book/{id}:
    get:
      summary: Get a specific book by ID
      description: Retrieve a specific book by its unique identifier.
      operationId: getBookById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the book to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: The book object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDTO'
        '404':
          description: Book not found
        '500':
          description: Internal server error
    put:
      summary: Update a book
      description: Update the details of an existing book.
      operationId: updateBook
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the book to update
          schema:
            type: integer
      requestBody:
        description: Updated book object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookDTO'
      responses:
        '200':
          description: Book updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDTO'
        '400':
          description: Invalid book data
        '404':
          description: Book not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a book
      description: Deletes a book by its unique identifier.
      operationId: deleteBook
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the book to delete
          schema:
            type: integer
      responses:
        '204':
          description: Book deleted successfully
        '404':
          description: Book not found
        '500':
          description: Internal server error
  /anime:
    get:
      summary: Get all anime
      description: Retrieve a list of all anime in the database, with optional pagination.
      operationId: getAllAnime
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size for pagination
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of anime
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnimeDTO'
        '500':
          description: Internal server error
    post:
      summary: Create a new anime
      description: Adds a new anime to the database.
      operationId: createAnime
      requestBody:
        description: Anime object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnimeDTO'
      responses:
        '201':
          description: Anime created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimeDTO'
        '400':
          description: Invalid anime data
        '500':
          description: Internal server error
  /anime/{id}:
    get:
      summary: Get a specific anime by ID
      description: Retrieve a specific anime by its unique identifier.
      operationId: getAnimeById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the anime to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: The anime object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimeDTO'
        '404':
          description: Anime not found
        '500':
          description: Internal server error
    put:
      summary: Update an anime
      description: Update the details of an existing anime.
      operationId: updateAnime
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the anime to update
          schema:
            type: integer
      requestBody:
        description: Updated anime object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnimeDTO'
      responses:
        '200':
          description: Anime updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimeDTO'
        '400':
          description: Invalid anime data
        '404':
          description: Anime not found
        '500':
          description: Internal server error
    delete:
      summary: Delete an anime
      description: Deletes an anime by its unique identifier.
      operationId: deleteAnime
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the anime to delete
          schema:
            type: integer
      responses:
        '204':
          description: Anime deleted successfully
        '404':
          description: Anime not found
        '500':
          description: Internal server error
components:
  schemas:
...
    ComicsDTO:
      $ref: './media-api-schematic.yaml#/components/schemas/ComicsDTO'
    BookDTO:
      $ref: './media-api-schematic.yaml#/components/schemas/BookDTO'
    AnimeDTO:
      $ref: './media-api-schematic.yaml#/components/schemas/AnimeDTO'